refactor: payment_request module
