--- packages/cryptoplease/lib/app/screens/authenticated/receive_flow/routes.dart
@@ -1,7 +1,0 @@
-import '../../../../features/payment_request/presentation/link_details/routes.dart';
-import '../../../../features/payment_request/presentation/link_request/routes.dart';
-
-const receiveFlowRoutes = [
-  linkRequestRoutes,
-  ...linkDetailsRoutes,
-];
