--- packages/cryptoplease/lib/features/payment_request/presentation/link_request/routes.dart
@@ -1,5 +1,0 @@
-import 'package:auto_route/auto_route.dart';
-
-import 'flow.dart';
-
-const linkRequestRoutes = AutoRoute<void>(page: LinkRequestFlowScreen);
