+++ packages/cryptoplease/lib/features/payment_request/src/bl/create_payment_request/state.dart
