+++ packages/cryptoplease/lib/features/payment_request/src/bl/repository.dart
@@ -6,7 +6,7 @@ import 'package:injectable/injectable.dart';
 import 'package:solana/solana.dart';
 import 'package:solana/solana_pay.dart';
 
-import '../../../data/db/db.dart';
+import '../../../../data/db/db.dart';
 import 'payment_request.dart';
 
 @injectable
