+++ packages/cryptoplease/lib/features/payment_request/src/presentation/link_details/share_link.dart
@@ -3,18 +3,18 @@ import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:share/share.dart';
 
-import '../../../../core/presentation/format_amount.dart';
-import '../../../../core/presentation/utils.dart';
-import '../../../../core/tokens/token_list.dart';
-import '../../../../di.dart';
-import '../../../../l10n/l10n.dart';
-import '../../../../ui/app_bar.dart';
-import '../../../../ui/button.dart';
-import '../../../../ui/colors.dart';
-import '../../../../ui/content_padding.dart';
-import '../../../../ui/share_message/share_message_bubble.dart';
-import '../../../../ui/share_message/share_message_header.dart';
-import '../../../../ui/theme.dart';
+import '../../../../../core/presentation/format_amount.dart';
+import '../../../../../core/presentation/utils.dart';
+import '../../../../../core/tokens/token_list.dart';
+import '../../../../../di.dart';
+import '../../../../../l10n/l10n.dart';
+import '../../../../../ui/app_bar.dart';
+import '../../../../../ui/button.dart';
+import '../../../../../ui/colors.dart';
+import '../../../../../ui/content_padding.dart';
+import '../../../../../ui/share_message/share_message_bubble.dart';
+import '../../../../../ui/share_message/share_message_header.dart';
+import '../../../../../ui/theme.dart';
 import '../../bl/payment_request.dart';
 
 class SharePaymentRequestLinkScreen extends StatelessWidget {
