+++ packages/cryptoplease/lib/features/payment_request/src/presentation/link_details/share_qr_code.dart
