+++ packages/cryptoplease/lib/features/payment_request/src/presentation/link_request/request_note_screen.dart
@@ -3,16 +3,16 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
-import '../../../../core/amount.dart';
-import '../../../../core/presentation/format_amount.dart';
-import '../../../../l10n/device_locale.dart';
-import '../../../../l10n/l10n.dart';
-import '../../../../ui/app_bar.dart';
-import '../../../../ui/bordered_row.dart';
-import '../../../../ui/bottom_button.dart';
-import '../../../../ui/colors.dart';
-import '../../../../ui/content_padding.dart';
-import '../../../../ui/theme.dart';
+import '../../../../../core/amount.dart';
+import '../../../../../core/presentation/format_amount.dart';
+import '../../../../../l10n/device_locale.dart';
+import '../../../../../l10n/l10n.dart';
+import '../../../../../ui/app_bar.dart';
+import '../../../../../ui/bordered_row.dart';
+import '../../../../../ui/bottom_button.dart';
+import '../../../../../ui/colors.dart';
+import '../../../../../ui/content_padding.dart';
+import '../../../../../ui/theme.dart';
 
 abstract class NoteSetter {
   void onNoteSubmitted(String name);
