+++ packages/cryptoplease/lib/features/payment_request/src/presentation/link_request/routes.dart
@@ -1,0 +1,7 @@
+import 'package:auto_route/auto_route.dart';
+
+import 'flow.dart';
+
+const linkRequestRoutes = [
+  AutoRoute<void>(page: LinkRequestFlowScreen),
+];
